API

Documentation

data Message

Constructors

Message

Fields

name :: String

message :: String

data ResponseData

We will also define a simple data type for returning data from a REST call, again with nothing special or particular in the response, but instead merely as a demonstration.

Constructors

ResponseData

Fields

response :: String

type FileAPI = ("postFile" :> (ReqBody '[JSON] Message :> Post '[JSON] Bool)) :<|> ("getFile" :> (QueryParam "name" String :> Get '[JSON] [Message]))

Next we will define the API for the REST service. This is defined as a 'type' using a special syntax from the Servant Library. A REST endpoint is defined by chaining together a series of elements in the format `A :> B :> C`. A set of rest endpoints are chained in the format `X :| Y :| Z`. We define a set of endpoints to demonstrate functionality as described int he README.md file below.

Note in the API below that we can mix GET and Post methods. The type of call is determined by the last element in the :> chain. If the method is Get, then the set of QueryParams determine the attributes of the Get call. If the method is Post, then there will be a single ReqBody element that defines the type being transmitted. The return type for each method is noted in the last element in the :> chain.

type DirectoryAPI = ("postDirectory" :> (ReqBody '[JSON] Message :> Post '[JSON] Bool)) :<|> ("getDirectory" :> (QueryParam "name" String :> Get '[JSON] [Message]))

Orphan instances

ToBSON [Char]

Methods

toBSON :: [Char] -> Document

FromBSON [Char]

Methods

fromBSON :: Document -> Maybe [Char]