ClientAPI

Documentation

restAPI :: Proxy API

postFile :: Message -> ClientM Bool

The following provides the implementations of these types Note that the order of the functions must match the endpoints in the type API from UseHaskell.hs

The function type of the interface here. Each function matches one of the endpoints in type API from UseHaskellAPI.hs

getFile :: Maybe String -> ClientM [Message]